requirements.txt
python-dotenv==0.21.0
requests==2.27.1
networkx==2.8.7
matplotlib==3.6.2